requirements.txt
# Core Dependencies
pandas==1.5.3
numpy==1.24.4

# AI & ML
langchain==0.1.12
langchain-google-genai==0.0.1
google-generativeai==0.3.2
crewai==0.28.0
crewai-tools==0.1.5
chroma-hnswlib==0.7.3
langchain-community==0.0.28

# Vector Database
chromadb==0.4.23

# UI
streamlit==1.24.1

# Data Processing
pymupdf==1.22.5
pytesseract==0.3.10
python-dotenv==1.0.0

# Evaluation
scikit-learn==1.2.2
matplotlib==3.7.2
seaborn==0.12.2

# Deployment
docker==6.1.3

# ONNX runtime (required for some models)
onnxruntime==1.16.3
